requirements.txt
asgiref==3.7.2
attrs==23.2.0
Babel==2.10.3
blinker==1.7.0
cachetools==5.5.2
certifi==2023.11.17
chardet==5.2.0
click==8.1.6
colorama==0.4.6
configobj==5.0.8
coreapi==2.3.3
coreschema==0.0.4
crispy-bootstrap5==2023.10
cryptography==41.0.7 
distlib==0.3.8
distro==1.9.0
Django==4.2.11
django-cors-headers==4.3.1
django-crispy-forms==2
django-filter==23.5
django-guardian==2.4.0
django-maintenance-mode==0.16.1
djangorestframework==3.14.0
djangorestframework_simplejwt==5.5.0
filelock==3.13.1
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-recaptcha-enterprise==1.28.2
googleapis-common-protos==1.70.0
grpcio==1.73.1
grpcio-status==1.73.1
httplib2==0.20.4
idna==3.6
itypes==1.2.0
Jinja2==3.1.2
jsonpatch==1.32
jsonpointer==2.0
jsonschema==4.10.3
lazr.restfulclient==0.14.6
lazr.uri==1.0.6 
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
netaddr==0.8.0
netifaces==0.11.0
oauthlib==3.2.2
olefile==0.46
pexpect==4.9.0
# pillow==10.2.0
platformdirs==4.2.0
proto-plus==1.26.1
protobuf==6.31.1
psycopg2==2.9.10
ptyprocess==0.7.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
Pygments==2.17.2
PyJWT==2.7.0
pyparsing==3.1.1
pyrsistent==0.20.0
pyserial==3.5
python-dateutil==2.8.2
pytz==2024.1
pyxdg==0.28
PyYAML==6.0.1
requests==2.31.0
rich==13.7.1
rsa==4.9.1
setuptools==68.1.2
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.10.0
uritemplate==4.1.1
urllib3==2.0.7
virtualenv==20.25.0 
wheel==0.42.0
whitenoise==6.0.0
django-tailwind
pillow
